+++ components/index.js
@@ -69,7 +69,10 @@ export SLDSIcon from './icon';
 export Icon from './icon';
 
 export SLDSButtonIcon from './icon/button-icon';
-export ButtonIcon from './icon';
+export ButtonIcon from './icon/button-icon';
+
+export SLDSInlineEdit from './forms/input/inline';
+export InlineEdit from './forms/input/inline';
 
 export SLDSInput from './forms/input';
 export Input from './forms/input';
@@ -97,6 +100,9 @@ export PageHeader from './page-header';
 
 export SLDSPopoverTooltip from './popover-tooltip';
 export PopoverTooltip from './popover-tooltip';
+
+export SLDSSpinner from './spinner';
+export Spinner from './spinner';
 
 export SLDSTimepicker from './time-picker';
 export Timepicker from './time-picker';