+++ stories/forms/input/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { storiesOf, action } from '@kadira/storybook';
 
-import { FORMS_INPUT } from '../../utilities/constants';
-import Input from '../../components/forms/input';
+import { FORMS_INPUT } from '../../../utilities/constants';
+import Input from '../../../components/forms/input';
 
 const iconClicked = action;
 
@@ -10,14 +10,12 @@ storiesOf(FORMS_INPUT, module)
 	.addDecorator(getStory => <div className="slds-p-around--medium">{getStory()}</div>)
 	.add('Standard Input', () => (
 		<Input
-			id="unique-id-1"
 			label="Input Label"
 			placeholder="Placeholder Text"
 		/>
 	))
 	.add('Assistive Text, No Label', () => (
 		<Input
-			id="unique-id-1"
 			assistiveText="Assistive Text"
 			placeholder="Placeholder Text"
 		/>
@@ -35,7 +33,6 @@ storiesOf(FORMS_INPUT, module)
 	))
 	.add('Input with Right Clickable Icon', () => (
 		<Input
-			id="unique-id-2"
 			label="Input Label"
 			iconName="close"
 			iconCategory="utility"
@@ -56,7 +53,6 @@ storiesOf(FORMS_INPUT, module)
 	))
 	.add('Read Only Input', () => (
 		<Input
-			id="unique-id-3"
 			label="Input Label"
 			readOnly
 			value="Read Only Value"
@@ -64,11 +60,9 @@ storiesOf(FORMS_INPUT, module)
 	))
 	.add('Required Input in Error State', () => (
 		<Input
-			id="unique-id-4"
 			label="Input Label"
 			required
 			errorText="Error Message"
 			placeholder="Placeholder Text"
 		/>
 	));
-