+++ tests/input/input.test.jsx
@@ -1,3 +1,6 @@
+/* eslint-env mocha */
+/* eslint-disable prefer-arrow-callback */
+
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { expect } from 'chai';
@@ -14,7 +17,6 @@ const {
 
 describe('SLDS INPUT **************************************************', () => {
 	const defaultProps = {
-		id: 'unique-id-2',
 		placeholder: 'Placeholder Text'
 	};