+++ RELEASENOTES.md
@@ -2,6 +2,21 @@
 # React Components: design-system-react
 # Release notes
 
+## Release 0.0.32
+- Added containerClassName prop to Spinner
+
+## Release 0.0.31
+
+**MAJOR CHANGES**
+- Added new Spinner component
+- Updated the ContextBar component
+- Added a highlighting utility
+- Added a custom cell type to DataTable using the highlighter
+
+**OTHER**
+- Remove isRequired for iconCategory in ButtonGroup
+- Additional bug fixes and code cleanup
+
 ## Release 0.0.29
 
 **MAJOR CHANGES**